refactor(embassy)!: extract the STM32 module into its own crate
