+++ src/riot-rs-stm32/src/usb.rs
@@ -1,14 +1,12 @@
 use embassy_stm32::{bind_interrupts, peripherals, usb, usb::Driver};
 
-use crate::arch;
-
 bind_interrupts!(struct Irqs {
     USB_LP => usb::InterruptHandler<peripherals::USB>;
 });
 
 pub type UsbDriver = Driver<'static, peripherals::USB>;
 
-pub fn driver(peripherals: &mut arch::OptionalPeripherals) -> UsbDriver {
+pub fn driver(peripherals: &mut crate::OptionalPeripherals) -> UsbDriver {
     let usb = peripherals.USB.take().unwrap();
     let dp = peripherals.PA12.take().unwrap();
     let dm = peripherals.PA11.take().unwrap();
